adds Gamry EIS script
Gamry EIS script handles electrochemical impedance spectroscopy and polarization curve data. If 'EIS Summary' mode is selected, Gamry type files are detected in the selected directory and then turned into pandas dataframes. the high frequency intercept (HFR) is interpolated or extracted (depending on quality of the data set) and then the pol curve, HFR, and HFR-free data is plotted and stored in a newly created 'Processed EIS' folder. Plots are created and stored which show the quality of the fit and a .csv file with all pertinent data is created. If non-faradaic impedance is performed, then a fitting of the catalyst layer resistance is attempted. 'EIS Manual fit' mode creates .csv type files for all EIS data files so that one can manually fit the data if desired.